+++ lib/src/flexible_bottom_sheet.dart
@@ -107,6 +107,7 @@ class FlexibleBottomSheet extends StatefulWidget {
   final VoidCallback? onDismiss;
   final Color? keyboardBarrierColor;
   final Color? bottomSheetColor;
+  final bool useRootScaffold;
 
   FlexibleBottomSheet({
     Key? key,
@@ -127,6 +128,7 @@ class FlexibleBottomSheet extends StatefulWidget {
     this.keyboardBarrierColor,
     this.bottomSheetColor,
     this.draggableScrollableController,
+    this.useRootScaffold = true,
   })  : assert(minHeight >= 0 && minHeight <= 1),
         assert(maxHeight > 0 && maxHeight <= 1),
         assert(maxHeight > minHeight),
@@ -152,6 +154,7 @@ class FlexibleBottomSheet extends StatefulWidget {
     Decoration? decoration,
     Color? keyboardBarrierColor,
     Color? bottomSheetColor,
+    bool useRootScaffold = true,
   }) : this(
           key: key,
           maxHeight: maxHeight,
@@ -170,6 +173,7 @@ class FlexibleBottomSheet extends StatefulWidget {
           decoration: decoration,
           keyboardBarrierColor: keyboardBarrierColor,
           bottomSheetColor: bottomSheetColor,
+          useRootScaffold: useRootScaffold,
         );
 
   @override
@@ -350,11 +354,12 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
                 );
               }
             },
-            child: Scaffold(
+            child: _RegisterScaffold(
+              useRootScaffold: widget.useRootScaffold,
               backgroundColor: widget.bottomSheetColor ??
                   Theme.of(context).bottomSheetTheme.backgroundColor ??
                   Theme.of(context).colorScheme.background,
-              body: _Content(
+              child: _Content(
                 builder: widget.builder,
                 decoration: widget.decoration,
                 bodyBuilder: widget.bodyBuilder,
@@ -376,6 +381,33 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
         },
       ),
     );
+  }
+}
+
+/// Register [Scaffold] for [FlexibleBottomSheet].
+class _RegisterScaffold extends StatelessWidget {
+  final bool useRootScaffold;
+  final Widget child;
+  final Color backgroundColor;
+
+  const _RegisterScaffold({
+    required this.useRootScaffold,
+    required this.child,
+    required this.backgroundColor,
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return useRootScaffold
+        ? Scaffold(
+            backgroundColor: backgroundColor,
+            body: child,
+          )
+        : ColoredBox(
+            color: backgroundColor,
+            child: child,
+          );
   }
 }