+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased, estimated N.N.N]
 ## 3.2.0
+### Added
+* Added a new parameter that controls the display of the snackbar
 ### Fixed
 * Added Force widget rebuild to change bottomSheetOffset value